+++ HardCoreEnemies/Enemies/DamageOutput.lua
@@ -1,4 +1,4 @@
-local Folder = require("Functions.Function")
+local Folder = require("Functions.Damages")
 
 Folder.BitersDamage()
 Folder.PentapodDamage()
