+++ HardCoreEnemies/Functions/Damages.lua
