+++ HardCoreEnemies/data.lua
@@ -1,3 +1,3 @@
 require("Enemies.DamageOutput")
 require("Enemies.ObjectResitances")
-require("Functions.Function")
+require("HardCoreEnemies.Functions.Damages")
